\name{Cmeasure}
\alias{Cmeasure}
\title{ Calculate the C-Measure}
\arguments{
\item{x}{Vektor der Punkte in Eingaberaum}
\item{y}{Vektor der Punkte in Ausgaberaum}
\item{k}{Anzahl der Nachbarn in der Naehe}
\item{method}{es wird nur Minimal Pathlength und Minimal Wiring implementiert}
\item{p}{fuer Minimal Wiring 'wiring', Ausgaberaum, default=1}
\item{distance}{Distanzmass in Nachbarnschaftberechnung}
}
\value{cmeasure    the calculated c measure}
\description{
 Calculate the C-Measure}
\author{
Becker, Seibert,Roth,Zhang 07/2015}
