\name{ForceApproachErrorFL}
\alias{ForceApproachErrorFL}
\title{ Berchnet den Force Approach Error }
\arguments{
\item{VectorOfInputDists(1:n2)}{dissimilarities in Input Space between the n data points in vector form as produced by squareform(Dists(1:n,1:n))}
\item{VectorOfOutputDists(1:n2)}{dissimilarities in Input Space between the n data points in vector form as produced by squareform(Dists(1:n,1:n))}
}
\value{ForceApproachError                                   Force Approach Error}
\description{
 Berchnet den Force Approach Error }
\author{
FL}
