\name{ForceApproachQ}
\alias{ForceApproachQ}
\title{ Force Approach Error Q }
\arguments{
\item{VectorOfInputDists(1:n2)}{dissimilarities in Input Space between the n data points in vector form as produced by squareform(Dists(1:n,1:n))}
\item{VectorOfOutputDists(1:n2)}{dissimilarities in Input Space between the n data points in vector form as produced by squareform(Dists(1:n,1:n))}
\item{k}{order of the Nearest Neighborhood (kNN)}
}
\value{Q                                   Forced appraoch Error}
\description{
 Force Approach Error Q }
\author{
MT}
