\name{LCMC}
\alias{LCMC}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
LCMC
}
\description{
Implementation of LCMC
}
\usage{
LCMC(InputDist, OutputDist, k)
}
\arguments{
  \item{InputDist}{
Distance Matrix of Data
}
  \item{OutputDist}{
Distance Matrix of ProjectedPoints
}
  \item{k}{ An integer, giving the 
	number of nearest neighbors to i
}
}

\value{A list of:
\item{MK}{See references}
\item{MKadj}{See references}
}
\references{
TODO
%% ~put references to the literature/web site here ~
}
\author{
Marta Lemanczyk
}