\name{LocallyLinearEmbedding}
\alias{LocallyLinearEmbedding}
\title{ Locally Linear Embedding as introduced in 2000 by Roweis, Saul and Lawrence. projection=LocallyLinearEmbedding(Data,KNN=10)}
\arguments{
\item{Data[1:n,1:d]}{array of data: n cases in rows, d variables in columns}
}
\value{of the dimension chosen by cmdscale and OutputDimension}
\description{
 Locally Linear Embedding as introduced in 2000 by Roweis, Saul and Lawrence. projection=LocallyLinearEmbedding(Data,KNN=10)}
\author{
MT 06/2015}
