\name{MeanRelativeRankError}
\alias{MeanRelativeRankError}
\title{Berechnet den Mean RelativeRank Error}
\arguments{
\item{VectorOfInputDist}{Vektor mit den Distanzen im Eingaberaum}
\item{VectorOfOutputDists}{Vektor mit den Distanzen im Ausgaberaum}
\item{k(optional)}{k für die Bestimmung der k nächsten NachbarnU}
}
\value{
\item{MRRE1}{Mean Relative Rank Error 1}
\item{MRRE2}{Mean Relative Rank Error 2}
\item{C}{Mit 1/C wurden die Werte normiert}
}
\description{
 Berechnet den Mean RelativeRank Error}
\author{Florian Lerch}
