% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_add_projections.R
\name{merge_add_projections}
\alias{merge_add_projections}
\alias{+.projections}
\title{Add data of different projections objects}
\usage{
merge_add_projections(x)

\method{+}{projections}(a, b)
}
\arguments{
\item{x}{A `list` of `projections` objects to be added.}

\item{a}{A  `projections` object.}

\item{b}{A  `projections` object.}
}
\description{
This function adds counts from several `projections` objects, making sure
that they all use the same dates, adding rows of '0' where
needed. Simulations (columns) are recycled when needed if some objects have
less simulations than others. The same operation is implemented by the `+`
operator.
}
\examples{


}
\author{
Thibaut Jombart
}
