% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_projections.R
\name{merge_projections}
\alias{merge_projections}
\title{Merge a list of projections objects}
\usage{
merge_projections(x)
}
\arguments{
\item{x}{a `list` of `projections` objects to be merged}
}
\description{
This function merges `projections` objects.
}
