% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.projections.R
\name{print.projections}
\alias{print.projections}
\title{Print method for projections objects}
\usage{
\method{print}{projections}(x, ...)
}
\arguments{
\item{x}{A \code{projections} object.}

\item{...}{further parameters to be passed to other methods (currently not
used)}
}
\description{
This method prints the content of \code{projections} objects.
}
\author{
Thibaut Jombart (\email{thibautjombart@gmail.com})
}
