% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{project}
\alias{project}
\title{Project future incidence}
\usage{
project(x, R, si, n_sim = 100, n_days = 7)
}
\arguments{
\item{x}{An \code{incidence} object containing daily incidence; other time
intervals will trigger an error.}

\item{R}{A vector of numbers representing plausible reproduction numbers; for
instance, these can be samples from a posterior distribution using the
\code{EpiEstim} package.}

\item{si}{A function computing the serial interval, or a \code{numeric}
vector providing its mass function. For functions, we strongly recommend
using the RECON package \code{distcrete} to obtain such distribution (see
example).}

\item{n_sim}{The number of epicurves to simulate. Defaults to 100.}

\item{n_days}{The number of days to run simulations for. Defaults to 14.}
}
\description{
This function simulates future incidence based on past incidence data, a
selection of plausible reproduction numbers (R), and the distribution of the
serial interval (time from primary onset to secondary onset).
}
\examples{

if (require(distcrete) && require(incidence)) {

## simulate basic epicurve
dat <- c(0, 2, 2, 3, 3, 5, 5, 5, 6, 6, 6, 6)
x <- incidence(dat)
plot(x)


## example with a function for SI

si <- distcrete("gamma", interval = 1L,
                 shape = 5,
                 scale = 5)
barplot(si$d(0:100), main = "Serial Interval")


pred_1 <- project(x, 1.2, si, n_days = 20)
pred_1
my_col <- rgb(.1, .1, .8, .4)
matplot(pred1, type = "l", lty = 1, col = my_col)


## example with empirical serial interval
si <- c(0, 1, 2, 1, 0.5)
pred_2 <- project(x, 1.2, si, n_days = 30)
pred_2
matplot(pred2, type = "l", lty = 1, col = my_col)


}

}
