\name{ProjectedPoints2Bestmatches}
\alias{ProjectedPoints2Bestmatches}
\title{
Convert to Bestmatches
}
\description{
Converts cartesian coordinates to quadgrid coordinates  (= BestMatchingUnit coordinates)
}
\usage{
ProjectedPoints2Bestmatches(ProjectedPoints, Lines)
}
\arguments{
  \item{ProjectedPoints}{
A matrix containing cartesian coordinates
}
  \item{Lines}{
 Dimensions of overlying grid
}
}
\value{
Quadgrid coordinates
}
\author{
Michael Thrun
}