\name{Projections-package}
\Rdversion{1.1}
\alias{Projections-package}
\alias{Projections}
\alias{dbt.Projections-package}
\alias{dbt.Projections}
\docType{package}
\title{
Projections
}
\description{This package is designed for Projections, like PCA, ICA, MDS.}
\author{
Michael Thrun, Florian Lerch, Felix Pape
}

