\name{SpearmanError}
\alias{SpearmanError}
\title{ Calculates the error of a projection with spearman's rank correlation coefficient }
\arguments{
\item{VectorOfInputDists(1:n2)}{dissimilarities in Input Space between the n data points in vector form as produced by squareform(Dists(1:n,1:n))}
\item{VectorOfOutputDists(1:n2)}{dissimilarities in Input Space between the n data points in vector form as produced by squareform(Dists(1:n,1:n))}
}
\value{rho                                 rank correlation coefficient}
\description{
 Calculates the error of a projection with spearman's rank correlation coefficient }
\author{
Florian Lerch}
