% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{[.projections}
\alias{[.projections}
\alias{"subset.projections"}
\alias{"[.projections"}
\title{Subsetting 'projections' objects}
\usage{
\method{[}{projections}(x, i, j)
}
\arguments{
\item{x}{An projections object, generated by the function
\code{\link{project}}.}

\item{i}{a subset of dates to retain}

\item{j}{a subset of groups to retain}
}
\description{
Two functions can be used to subset projections objects. The operator "[" can
be used as for matrices, using the syntax \code{x[i,j]} where 'i' is a subset
of dates, and 'j' is a subset of simulations.
}
\seealso{
The \code{\link{project}} function to generate the 'projections'
  objects.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
