% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.projections}
\alias{summary.projections}
\title{Summary for projections objects}
\usage{
\method{summary}{projections}(
  object,
  quantiles = c(0.025, 0.25, 0.5, 0.75, 0.975),
  mean = TRUE,
  sd = TRUE,
  min = TRUE,
  max = TRUE,
  ...
)
}
\arguments{
\item{object}{A `projections` object to summarise}

\item{mean}{a `logical` indicating of the mean should be computed}

\item{sd}{a `logical` indicating of the standard deviation should be computed}

\item{min}{a `logical` indicating of the minimum should be computed}

\item{max}{a `logical` indicating of the maximum should be computed}

\item{...}{only preesnt for compatibility with the generic}

\item{quantile}{A `numeric` vector indicating which quantiles should be
computed; ignored if `FALSE` or of length 0}
}
\description{
This method summarises predicted epidemic trajectories contained in a
`projections` object by days, deriving the mean, standard deviation, and
user-specified quantiles for each day.
}
\author{
Thibaut Jombart
}
