% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_add_projections.R
\name{merge_add_projections}
\alias{merge_add_projections}
\title{Add cases from different projections objects}
\usage{
merge_add_projections(x)
}
\arguments{
\item{x}{a `list` of `projections` objects to be merged}
}
\description{
This function adds cases from a list of `projections` objects with the same
number of simulations, but potentially different dates.
}
