% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_projections.R
\name{merge_projections}
\alias{merge_projections}
\title{Merge a list of projections objects}
\usage{
merge_projections(x)
}
\arguments{
\item{x}{A `list` of `projections` objects to be merged.}
}
\description{
This function merges `projections` objects, binding them by columns, making
sure that they all use the same dates, adding rows of '0' where needed.
}
\examples{

}
\author{
Thibaut Jombart
}
