% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.projections.R
\name{plot.projections}
\alias{plot.projections}
\title{Plot projections objects}
\usage{
\method{plot}{projections}(x, y = NULL, ...)
}
\arguments{
\item{x}{A \code{projections} object.}
}
\description{
This method is designed for plotting \code{projections} objects, output by
the function \code{\link{project}}.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
