% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.projections.R
\name{plot.projections}
\alias{plot.projections}
\alias{add_projections}
\title{Plot projections objects}
\usage{
\method{plot}{projections}(x, ...)

add_projections(p, x, quantiles = c(0.01, 0.05, 0.1, 0.5), boxplots = TRUE,
  palette = quantile_pal)
}
\arguments{
\item{x}{A \code{projections} object.}

\item{...}{Further arguments to be passed to other methods (not used).}

\item{p}{A previous incidence plot to which projections should be added.}

\item{quantiles}{A vector of quantiles to plot, automatically completed to be
symmetric around the median.}

\item{boxplots}{A logical indicating if boxplots should be drawn.}

\item{palette}{A color palette to be used for plotting the quantile lines;
defaults to \code{quantile_pal}.}
}
\description{
The \code{plot} method of \code{projections} objects (output by the function
\code{\link{project}}) shows quantiles of predicted incidence over time. The
function \code{add_projections} can be used to add a similar plot to an
existing \code{incidence} plot. This latter function is piping friendly (see
examples).
}
\examples{

if (require(outbreaks) &&
    require(distcrete) &&
    require(incidence) &&
    require(magrittr)) {

si <- distcrete("gamma", interval = 1L,
                 shape = 0.37,
                 scale = 41.4, w = 0)

i <- incidence(ebola_sim$linelist$date_of_onset)
plot(i)

## add projections after the first 100 days, over 60 days
set.seed(1)
proj <- project(x = i[1:100], R = 2.1, si = si, n_days = 60)

## plotting projections
plot(proj)

## adding them to incidence plot
plot(i) \%>\% add_projections(proj)
plot(i[1:160]) \%>\% add_projections(proj)

## same, with customised quantiles and colors
quantiles <- c(.001, .01, 0.05, .1, .2, .3, .4, .5)
pal <- colorRampPalette(c("#b3c6ff", "#00e64d", "#cc0066"))
plot(i[1:200]) \%>\%
  add_projections(proj, quantiles, palette = pal)

}

}
\seealso{
\code{\link{project}} to generate projections
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
