% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{[.projections}
\alias{[.projections}
\alias{"subset.projections"}
\alias{"[.projections"}
\alias{subset.projections}
\title{Subsetting 'projections' objects}
\usage{
\method{[}{projections}(x, i, j)

\method{subset}{projections}(x, ..., from = NULL, to = NULL,
  sim = TRUE)
}
\arguments{
\item{x}{An projections object, generated by the function
\code{\link{project}}.}

\item{i}{a subset of dates to retain}

\item{j}{a subset of groups to retain}

\item{...}{Further arguments passed to other methods (not used).}

\item{from}{The starting date; data strictly before this date are discarded.}

\item{to}{The ending date; data strictly after this date are discarded.}

\item{sim}{(optional) The simulations to retained, indicated as subsets of
the columns of x.}
}
\description{
Two functions can be used to subset projections objects. The operator "[" can
be used as for matrices, using the syntax \code{x[i,j]} where 'i' is a subset
of dates, and 'j' is a subset of simulations.
}
\seealso{
The \code{\link{project}} function to generate the 'projections'
  objects.
}
\author{
Thibaut Jombart \email{thibautjombart@gmail.com}
}
